26/10

+++ A-CTU/main/include/aux_ctu_hw.h
@@ -58,7 +58,7 @@ uint8_t low_power;
 uint8_t full_power;
 uint8_t or_gate;
 
-int FOD_counter;
+uint64_t FOD_counter;
 
 /* Semaphore used to protect against I2C reading simultaneously */
 SemaphoreHandle_t i2c_sem;
